only overwrite Arkime's config.ini with config.orig.ini config.ini doesn't already exist idaholab/Malcolm#311

+++ arkime/scripts/docker_entrypoint.sh
@@ -50,7 +50,8 @@ if ( [[ "$OPENSEARCH_PRIMARY" == "opensearch-remote" ]] || [[ "$OPENSEARCH_PRIMA
     OPENSEARCH_URL_FINAL="${PROTOCOL}${NEW_USER}:${NEW_PASSWORD}@${HOSTPORT}"
 fi
 
-if [[ -r "${ARKIME_DIR}"/etc/config.orig.ini ]]; then
+# iff config.ini does not exist but config.orig.ini does, use it as a basis and modify based on env. vars
+if [[ ! -f "${ARKIME_CONFIG_FILE}" ]] && [[ -r "${ARKIME_DIR}"/etc/config.orig.ini ]]; then
     cp "${ARKIME_DIR}"/etc/config.orig.ini "${ARKIME_CONFIG_FILE}"
 
     sed -i "s|^\(elasticsearch=\).*|\1"${OPENSEARCH_URL_FINAL}"|" "${ARKIME_CONFIG_FILE}"
